refactor: Move alwaysCallAfter decorator to separate module and update imports
