+++ addon/globalPlugins/remoteClient/alwaysCallAfter.py
@@ -1,0 +1,9 @@
+from functools import wraps
+
+
+def alwaysCallAfter(func):
+	"""Decorator to call a function after the current event loop has finished."""
+	@wraps(func)
+	def wrapper(*args, **kwargs):
+		wx.CallAfter(func, *args, **kwargs)
+	return wrapper
